Changed issue 38 fix: went back to 'int' coordinates, because that actually yields smaller (and probably faster) code. Also put scrollback up to 64k lines again.

+++ branches/0.3/term.h
@@ -79,7 +79,7 @@
 #define TSIZE (sizeof(TTYPE))
 
 typedef struct {
-  short y, x;
+  int y, x;
 } pos;
 
 typedef enum {
